unify Layer model and Db model
