+++ tipg/dbmodel.py
@@ -1,13 +1,82 @@
 """tipg.dbmodel: database events."""
 
-from typing import Any, Dict, List, Optional
+import re
+from typing import Any, Dict, List, Optional, Tuple, TypedDict, Union
 
-from buildpg import asyncpg
-from pydantic import BaseModel, Field
+from buildpg import RawDangerous as raw
+from buildpg import asyncpg, clauses
+from buildpg import funcs as pg_funcs
+from buildpg import logic, render
+from ciso8601 import parse_rfc3339
+from morecantile import Tile, TileMatrixSet
+from pydantic import BaseModel, Field, root_validator
+from pygeofilter.ast import AstType
 
+from tipg.errors import (
+    InvalidDatetime,
+    InvalidDatetimeColumnName,
+    InvalidGeometryColumnName,
+    InvalidLimit,
+    InvalidPropertyName,
+    MissingDatetimeColumn,
+)
+from tipg.filter.evaluate import to_filter
+from tipg.filter.filters import bbox_to_wkt
 from tipg.settings import TableSettings, TileSettings
 
 tile_settings = TileSettings()
+
+
+def debug_query(q, *p):
+    """Utility to print raw statement to use for debugging."""
+
+    qsub = re.sub(r"\$([0-9]+)", r"{\1}", q)
+
+    def quote_str(s):
+        """Quote strings."""
+
+        if s is None:
+            return "null"
+        elif isinstance(s, str):
+            return f"'{s}'"
+        else:
+            return s
+
+    p = [quote_str(s) for s in p]
+    print("DEBUG QUERY")
+    print(qsub.format(None, *p))
+
+
+# Links to geojson schema
+geojson_schema = {
+    "GEOMETRY": "https://geojson.org/schema/Geometry.json",
+    "POINT": "https://geojson.org/schema/Point.json",
+    "MULTIPOINT": "https://geojson.org/schema/MultiPoint.json",
+    "LINESTRING": "https://geojson.org/schema/LineString.json",
+    "MULTILINESTRING": "https://geojson.org/schema/MultiLineString.json",
+    "POLYGON": "https://geojson.org/schema/Polygon.json",
+    "MULTIPOLYGON": "https://geojson.org/schema/MultiPolygon.json",
+    "GEOMETRYCOLLECTION": "https://geojson.org/schema/GeometryCollection.json",
+}
+
+
+class Feature(TypedDict, total=False):
+    """Simple Feature model."""
+
+    type: str
+    # Geometry is either a dict or a str (wkt)
+    geometry: Optional[Union[Dict, str]]
+    properties: Optional[Dict]
+    id: Optional[Any]
+    bbox: Optional[List[float]]
+
+
+class FeatureCollection(TypedDict, total=False):
+    """Simple FeatureCollection model."""
+
+    type: str
+    features: List[Feature]
+    bbox: Optional[List[float]]
 
 
 class Column(BaseModel):
@@ -71,14 +140,17 @@ class Parameter(Column):
     default: Optional[str] = None
 
 
-class Table(BaseModel):
-    """Model for DB Table."""
+class Collection(BaseModel):
+    """Model for DB Table and Function."""
 
     type: str
     id: str
     table: str
     dbschema: str = Field(..., alias="schema")
+    title: Optional[str]
     description: Optional[str]
+    bounds: Optional[List[float]]  # TODO: Should be moved to Extent
+    crs: Optional[str]  # TODO: Should be moved to Extent
     properties: List[Column] = []
     id_column: Optional[str]
     geometry_columns: List[GeometryColumn] = []
@@ -89,6 +161,18 @@ class Table(BaseModel):
     minzoom: int = tile_settings.default_minzoom
     maxzoom: int = tile_settings.default_maxzoom
     default_tms: str = tile_settings.default_tms
+
+    @root_validator
+    def bounds_default(cls, values):
+        """Get default bounds from the first geometry columns."""
+        geoms = values.get("geometry_columns")
+        if geoms:
+            # Get the Extent of all the bounds
+            minx, miny, maxx, maxy = zip(*[geom.bounds for geom in geoms])
+            values["bounds"] = [min(minx), min(miny), max(maxx), max(maxy)]
+            values["crs"] = f"http://www.opengis.net/def/crs/EPSG/0/{geoms[0].srid}"
+
+        return values
 
     def get_datetime_column(self, name: Optional[str] = None) -> Optional[Column]:
         """Return the Column for either the passed in tstz column or the first tstz column."""
@@ -148,11 +232,549 @@ class Table(BaseModel):
 
         return None
 
+    def _select_no_geo(self, properties: Optional[List[str]], addid: bool = True):
+        nocomma = False
+        columns = self.columns(properties)
+        if columns:
+            sel = logic.as_sql_block(clauses.Select(columns))
+        else:
+            sel = logic.as_sql_block(raw("SELECT "))
+            nocomma = True
 
-Database = Dict[str, Dict[str, Any]]
+        if addid:
+            if self.id_column:
+                id_clause = logic.V(self.id_column).as_("tipg_id")
+            else:
+                id_clause = raw(" ROW_NUMBER () OVER () AS tipg_id ")
+            if nocomma:
+                sel = clauses.Clauses(sel, id_clause)
+            else:
+                sel = sel.comma(id_clause)
+
+        return logic.as_sql_block(sel)
+
+    def _select(
+        self,
+        properties: Optional[List[str]],
+        geometry_column: Optional[GeometryColumn],
+        bbox_only: Optional[bool],
+        simplify: Optional[float],
+        geom_as_wkt: bool = False,
+    ):
+        sel = self._select_no_geo(properties)
+
+        geom = self._geom(geometry_column, bbox_only, simplify)
+        if geom_as_wkt:
+            if geom:
+                sel = sel.comma(logic.Func("st_asewkt", geom).as_("tipg_geom"))
+            else:
+                sel = sel.comma(pg_funcs.cast(None, "text").as_("tipg_geom"))
+
+        else:
+            if geom:
+                sel = sel.comma(
+                    pg_funcs.cast(logic.Func("st_asgeojson", geom), "json").as_(
+                        "tipg_geom"
+                    )
+                )
+            else:
+                sel = sel.comma(pg_funcs.cast(None, "json").as_("tipg_geom"))
+
+        return sel
+
+    def _select_mvt(
+        self,
+        properties: Optional[List[str]],
+        geometry_column: GeometryColumn,
+        tms: TileMatrixSet,
+        tile: Tile,
+    ):
+        bbox = tms.xy_bounds(tile)
+        tms_srid = tms.crs.to_epsg()
+        if tms_srid:
+            transform_logic = logic.Func(
+                "st_transform",
+                logic.V(geometry_column.name),
+                pg_funcs.cast(tms_srid, "int"),
+            )
+        else:
+            tms_proj = str(tms.crs.to_proj4())
+            transform_logic = logic.Func(
+                "st_transform",
+                logic.V(geometry_column.name),
+                pg_funcs.cast(tms_proj, "text"),
+            )
+
+        sel = self._select_no_geo(properties, addid=False).comma(
+            logic.Func(
+                "st_asmvtgeom",
+                transform_logic,
+                logic.Func(
+                    "ST_Segmentize",
+                    logic.Func(
+                        "ST_MakeEnvelope",
+                        bbox.left,
+                        bbox.bottom,
+                        bbox.right,
+                        bbox.top,
+                    ),
+                    bbox.right - bbox.left,
+                ),
+                tile_settings.tile_resolution,
+                tile_settings.tile_buffer,
+            ).as_("geom")
+        )
+
+        return sel
+
+    def _select_count(self):
+        return clauses.Select(pg_funcs.count("*"))
+
+    def _from(self, function_parameters: Optional[Dict[str, str]]):
+        if self.type == "Function":
+            if not function_parameters:
+                return clauses.From(self.id) + raw("()")
+            params = []
+            for p in self.parameters:
+                if p.name in function_parameters:
+                    params.append(
+                        pg_funcs.cast(
+                            pg_funcs.cast(function_parameters[p.name], "text"),
+                            p.type,
+                        )
+                    )
+            return clauses.From(logic.Func(self.id, *params))
+        return clauses.From(self.id)
+
+    def _geom(
+        self,
+        geometry_column: Optional[GeometryColumn],
+        bbox_only: Optional[bool],
+        simplify: Optional[float],
+    ):
+        if geometry_column is None:
+            return None
+
+        g = logic.V(geometry_column.name)
+        g = pg_funcs.cast(g, "geometry")
+
+        if geometry_column.srid == 4326:
+            g = logic.Func("ST_Transform", g, pg_funcs.cast(4326, "int"))
+
+        if bbox_only:
+            g = logic.Func("ST_Envelope", g)
+        elif simplify:
+            g = logic.Func(
+                "ST_SnapToGrid",
+                logic.Func("ST_Simplify", g, simplify),
+                simplify,
+            )
+
+        return g
+
+    def _where(
+        self,
+        ids: Optional[List[str]] = None,
+        datetime: Optional[List[str]] = None,
+        bbox: Optional[List[float]] = None,
+        properties: Optional[List[Tuple[str, Any]]] = None,
+        cql: Optional[AstType] = None,
+        geom: Optional[str] = None,
+        dt: Optional[str] = None,
+        tile: Optional[Tile] = None,
+        tms: Optional[TileMatrixSet] = None,
+    ):
+        """Construct WHERE query."""
+        wheres = [logic.S(True)]
+
+        # `ids` filter
+        if ids is not None:
+            if len(ids) == 1:
+                wheres.append(
+                    logic.V(self.id_column)
+                    == pg_funcs.cast(
+                        pg_funcs.cast(ids[0], "text"), self.id_column_info.type
+                    )
+                )
+            else:
+                w = [
+                    logic.V(self.id_column)
+                    == logic.S(
+                        pg_funcs.cast(
+                            pg_funcs.cast(i, "text"), self.id_column_info.type
+                        )
+                    )
+                    for i in ids
+                ]
+                wheres.append(pg_funcs.OR(*w))
+
+        # `properties filter
+        if properties is not None:
+            w = []
+            for (prop, val) in properties:
+                col = self.get_column(prop)
+                if not col:
+                    raise InvalidPropertyName(f"Invalid property name: {prop}")
+
+                w.append(
+                    logic.V(col.name)
+                    == logic.S(pg_funcs.cast(pg_funcs.cast(val, "text"), col.type))
+                )
+
+            if w:
+                wheres.append(pg_funcs.AND(*w))
+
+        # `bbox` filter
+        geometry_column = self.get_geometry_column(geom)
+        if bbox is not None and geometry_column is not None:
+            wheres.append(
+                logic.Func(
+                    "ST_Intersects",
+                    logic.S(bbox_to_wkt(bbox)),
+                    logic.V(geometry_column.name),
+                )
+            )
+
+        # `datetime` filter
+        if datetime:
+            if not self.datetime_columns:
+                raise MissingDatetimeColumn(
+                    "Must have timestamp typed column to filter with datetime."
+                )
+
+            datetime_column = self.get_datetime_column(dt)
+            if not datetime_column:
+                raise InvalidDatetimeColumnName(f"Invalid Datetime Column: {dt}.")
+
+            wheres.append(self._datetime_filter_to_sql(datetime, datetime_column.name))
+
+        # `CQL` filter
+        if cql is not None:
+            wheres.append(to_filter(cql, [p.name for p in self.properties]))
+
+        if tile and tms and geometry_column:
+            tilebox = tms.bounds(tile)
+
+            wheres.append(
+                logic.Func(
+                    "ST_Intersects",
+                    logic.Func(
+                        "st_transform",
+                        logic.Func(
+                            "ST_Segmentize",
+                            logic.Func(
+                                "ST_MakeEnvelope",
+                                tilebox.left,
+                                tilebox.bottom,
+                                tilebox.right,
+                                tilebox.top,
+                                4326,
+                            ),
+                            tilebox.right - tilebox.left,
+                        ),
+                        pg_funcs.cast(geometry_column.srid, "int"),
+                    ),
+                    logic.V(geometry_column.name),
+                )
+            )
+
+        return clauses.Where(pg_funcs.AND(*wheres))
+
+    def _datetime_filter_to_sql(self, interval: List[str], dt_name: str):
+        if len(interval) == 1:
+            return logic.V(dt_name) == logic.S(
+                pg_funcs.cast(parse_rfc3339(interval[0]), "timestamptz")
+            )
+
+        else:
+            start = (
+                parse_rfc3339(interval[0]) if not interval[0] in ["..", ""] else None
+            )
+            end = parse_rfc3339(interval[1]) if not interval[1] in ["..", ""] else None
+
+            if start is None and end is None:
+                raise InvalidDatetime(
+                    "Double open-ended datetime intervals are not allowed."
+                )
+
+            if start is not None and end is not None and start > end:
+                raise InvalidDatetime("Start datetime cannot be before end datetime.")
+
+            if not start:
+                return logic.V(dt_name) <= logic.S(pg_funcs.cast(end, "timestamptz"))
+
+            elif not end:
+                return logic.V(dt_name) >= logic.S(pg_funcs.cast(start, "timestamptz"))
+
+            else:
+                return pg_funcs.AND(
+                    logic.V(dt_name) >= logic.S(pg_funcs.cast(start, "timestamptz")),
+                    logic.V(dt_name) < logic.S(pg_funcs.cast(end, "timestamptz")),
+                )
+
+    def _sortby(self, sortby: Optional[str]):
+        sorts = []
+        if sortby:
+            for s in sortby.strip().split(","):
+                parts = re.match(
+                    "^(?P<direction>[+-]?)(?P<column>.*)$", s
+                ).groupdict()  # type:ignore
+
+                direction = parts["direction"]
+                column = parts["column"].strip()
+                if self.get_column(column):
+                    if direction == "-":
+                        sorts.append(logic.V(column).desc())
+                    else:
+                        sorts.append(logic.V(column))
+                else:
+                    raise InvalidPropertyName(f"Property {column} does not exist.")
+
+        else:
+            if self.id_column is not None:
+                sorts.append(logic.V(self.id_column))
+            else:
+                sorts.append(logic.V(self.properties[0].name))
+
+        return clauses.OrderBy(*sorts)
+
+    async def _features_query(
+        self,
+        *,
+        pool: asyncpg.BuildPgPool,
+        ids_filter: Optional[List[str]] = None,
+        bbox_filter: Optional[List[float]] = None,
+        datetime_filter: Optional[List[str]] = None,
+        properties_filter: Optional[List[Tuple[str, str]]] = None,
+        cql_filter: Optional[AstType] = None,
+        sortby: Optional[str] = None,
+        properties: Optional[List[str]] = None,
+        geom: Optional[str] = None,
+        dt: Optional[str] = None,
+        limit: Optional[int] = None,
+        offset: Optional[int] = None,
+        bbox_only: Optional[bool] = None,
+        simplify: Optional[float] = None,
+        geom_as_wkt: bool = False,
+        function_parameters: Optional[Dict[str, str]],
+    ):
+        """Build Features query."""
+        c = clauses.Clauses(
+            self._select(
+                properties=properties,
+                geometry_column=self.get_geometry_column(geom),
+                bbox_only=bbox_only,
+                simplify=simplify,
+                geom_as_wkt=geom_as_wkt,
+            ),
+            self._from(function_parameters),
+            self._where(
+                ids=ids_filter,
+                datetime=datetime_filter,
+                bbox=bbox_filter,
+                properties=properties_filter,
+                cql=cql_filter,
+                geom=geom,
+                dt=dt,
+            ),
+            self._sortby(sortby),
+            clauses.Limit(limit or 10),
+            clauses.Offset(offset or 0),
+        )
+
+        q, p = render(":c", c=c)
+        async with pool.acquire() as conn:
+            for r in await conn.fetch(q, *p):
+                props = dict(r)
+                g = props.pop("tipg_geom")
+                id = props.pop("tipg_id")
+                feature = Feature(type="Feature", geometry=g, id=id, properties=props)
+                yield feature
+
+    async def _features_count_query(
+        self,
+        *,
+        pool: asyncpg.BuildPgPool,
+        ids_filter: Optional[List[str]] = None,
+        bbox_filter: Optional[List[float]] = None,
+        datetime_filter: Optional[List[str]] = None,
+        properties_filter: Optional[List[Tuple[str, str]]] = None,
+        cql_filter: Optional[AstType] = None,
+        geom: Optional[str] = None,
+        dt: Optional[str] = None,
+        function_parameters: Optional[Dict[str, str]],
+    ) -> int:
+        """Build features COUNT query."""
+        c = clauses.Clauses(
+            self._select_count(),
+            self._from(function_parameters),
+            self._where(
+                ids=ids_filter,
+                datetime=datetime_filter,
+                bbox=bbox_filter,
+                properties=properties_filter,
+                cql=cql_filter,
+                geom=geom,
+                dt=dt,
+            ),
+        )
+
+        q, p = render(":c", c=c)
+        async with pool.acquire() as conn:
+            count = await conn.fetchval(q, *p)
+            return count
+
+    async def features(
+        self,
+        pool: asyncpg.BuildPgPool,
+        *,
+        ids_filter: Optional[List[str]] = None,
+        bbox_filter: Optional[List[float]] = None,
+        datetime_filter: Optional[List[str]] = None,
+        properties_filter: Optional[List[Tuple[str, str]]] = None,
+        cql_filter: Optional[AstType] = None,
+        sortby: Optional[str] = None,
+        properties: Optional[List[str]] = None,
+        geom: Optional[str] = None,
+        dt: Optional[str] = None,
+        limit: Optional[int] = None,
+        offset: Optional[int] = None,
+        bbox_only: Optional[bool] = None,
+        simplify: Optional[float] = None,
+        geom_as_wkt: bool = False,
+        function_parameters: Dict[str, str] = {},
+    ) -> Tuple[FeatureCollection, int]:
+        """Build and run Pg query."""
+        if geom and geom.lower() != "none" and not self.get_geometry_column(geom):
+            raise InvalidGeometryColumnName(f"Invalid Geometry Column: {geom}.")
+
+        if limit and limit > tile_settings.max_features_per_tile:
+            raise
+
+        count = await self._features_count_query(
+            pool=pool,
+            ids_filter=ids_filter,
+            datetime_filter=datetime_filter,
+            bbox_filter=bbox_filter,
+            properties_filter=properties_filter,
+            function_parameters=function_parameters,
+            cql_filter=cql_filter,
+            geom=geom,
+            dt=dt,
+        )
+
+        features = [
+            f
+            async for f in self._features_query(
+                pool=pool,
+                ids_filter=ids_filter,
+                datetime_filter=datetime_filter,
+                bbox_filter=bbox_filter,
+                properties_filter=properties_filter,
+                cql_filter=cql_filter,
+                sortby=sortby,
+                properties=properties,
+                geom=geom,
+                dt=dt,
+                limit=limit,
+                offset=offset,
+                bbox_only=bbox_only,
+                simplify=simplify,
+                geom_as_wkt=geom_as_wkt,
+                function_parameters=function_parameters,
+            )
+        ]
+
+        return (
+            FeatureCollection(type="FeatureCollection", features=features),
+            count,
+        )
+
+    async def get_tile(
+        self,
+        *,
+        pool: asyncpg.BuildPgPool,
+        tms: TileMatrixSet,
+        tile: Tile,
+        ids_filter: Optional[List[str]] = None,
+        bbox_filter: Optional[List[float]] = None,
+        datetime_filter: Optional[List[str]] = None,
+        properties_filter: Optional[List[Tuple[str, str]]] = None,
+        function_parameters: Optional[Dict[str, str]] = None,
+        cql_filter: Optional[AstType] = None,
+        sortby: Optional[str] = None,
+        properties: Optional[List[str]] = None,
+        geom: Optional[str] = None,
+        dt: Optional[str] = None,
+        limit: Optional[int] = None,
+    ):
+        """Build query to get Vector Tile."""
+        geometry_column = self.get_geometry_column(geom)
+        if not geometry_column:
+            raise InvalidGeometryColumnName
+
+        if limit and limit > tile_settings.max_features_per_tile:
+            raise InvalidLimit
+
+        c = clauses.Clauses(
+            self._select_mvt(
+                properties=properties,
+                geometry_column=geometry_column,
+                tms=tms,
+                tile=tile,
+            ),
+            self._from(function_parameters),
+            self._where(
+                ids=ids_filter,
+                datetime=datetime_filter,
+                bbox=bbox_filter,
+                properties=properties_filter,
+                cql=cql_filter,
+                geom=geom,
+                dt=dt,
+                tms=tms,
+                tile=tile,
+            ),
+            clauses.Limit(limit or 10),
+        )
+
+        q, p = render(
+            """
+            WITH
+            t AS (:c)
+            SELECT ST_AsMVT(t.*) FROM t
+            """,
+            c=c,
+        )
+        # debug_query(q, *p)
+
+        async with pool.acquire() as conn:
+            return await conn.fetchval(q, *p)
+
+    @property
+    def queryables(self) -> Dict:
+        """Return the queryables."""
+        if self.geometry_columns:
+            geoms = {
+                col.name: {"$ref": geojson_schema.get(col.geometry_type.upper(), "")}
+                for col in self.geometry_columns
+            }
+        else:
+            geoms = {}
+
+        props = {
+            col.name: {"name": col.name, "type": col.json_type}
+            for col in self.properties
+            if col.name not in geoms
+        }
+
+        return {**geoms, **props}
 
 
-async def get_table_index(
+Database = Dict[str, Collection]
+
+
+async def get_collection_index(
     db_pool: asyncpg.BuildPgPool,
     schemas: Optional[List[str]] = ["public"],
     tables: Optional[List[str]] = None,
@@ -160,7 +782,7 @@ async def get_table_index(
     functions: Optional[List[str]] = None,
     spatial: bool = True,
 ) -> Database:
-    """Fetch Table index."""
+    """Fetch Table and Functions index."""
 
     query = """
         WITH table_columns AS (
@@ -446,19 +1068,19 @@ async def get_table_index(
             if not geometry_column and geometry_columns:
                 geometry_column = geometry_columns[0]
 
-            catalog[id] = {
-                "type": table["entity"],
-                "id": id,
-                "table": table["tbl"],
-                "schema": table["dbschema"],
-                "description": table["description"],
-                "id_column": id_column,
-                "geometry_columns": geometry_columns,
-                "datetime_columns": datetime_columns,
-                "properties": properties,
-                "datetime_column": datetime_column,
-                "geometry_column": geometry_column,
-                "parameters": table["parameters"],
-            }
+            catalog[id] = Collection(
+                type=table["entity"],
+                id=id,
+                table=table["tbl"],
+                schema=table["dbschema"],
+                description=table["description"],
+                id_column=id_column,
+                geometry_columns=geometry_columns,
+                datetime_columns=datetime_columns,
+                properties=properties,
+                datetime_column=datetime_column,
+                geometry_column=geometry_column,
+                parameters=table["parameters"],
+            )
 
         return catalog
